moved each EntityTable into db.tables package
